Move Model classes to model package and utils methods to utils package
