+++ app/src/main/java/com/omelan/cofi/pages/settings/licenses/DependencyItem.kt
@@ -14,66 +14,12 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.*
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
+import com.omelan.cofi.model.Dependency
+import com.omelan.cofi.model.License
 import com.omelan.cofi.pages.settings.settingsItemModifier
 import com.omelan.cofi.ui.Spacing
 import com.omelan.cofi.utils.URL_ANNOTATION
 import com.omelan.cofi.utils.appendLink
-import org.json.JSONArray
-import org.json.JSONObject
-
-data class License(val license: String, val license_url: String)
-
-data class Dependency(
-    val project: String,
-    val description: String,
-    val version: String,
-    val developers: List<String>,
-    val url: String?,
-    val year: String?,
-    val licenses: List<License>,
-)
-
-private fun JSONArray.toAuthorList(): List<String> {
-    val list = mutableListOf<String>()
-    (0 until this.length()).forEach {
-        list.add(this.getString(it))
-    }
-    return if (list.isEmpty()) listOf("Original author or authors") else list
-}
-
-private fun JSONArray.toLicensesList(): List<License> {
-    val list = mutableListOf<License>()
-    (0 until this.length()).forEach {
-        val jsonObject = this.getJSONObject(it)
-        list.add(
-            License(
-                license = jsonObject.getString("license"),
-                license_url = jsonObject.getString("license_url")
-            )
-        )
-    }
-    return list
-}
-
-fun String.parseJsonToDependencyList(): List<Dependency> {
-    val jsonArray = JSONArray(this)
-    val dependencyList = mutableListOf<Dependency>()
-    (0 until jsonArray.length()).forEach {
-        val jsonObject = jsonArray[it] as JSONObject
-        dependencyList.add(
-            Dependency(
-                project = jsonObject.getString("project"),
-                description = jsonObject.getString("description"),
-                version = jsonObject.getString("version"),
-                developers = jsonObject.getJSONArray("developers").toAuthorList(),
-                url = jsonObject.getString("url"),
-                year = jsonObject.getString("year"),
-                licenses = jsonObject.getJSONArray("licenses").toLicensesList(),
-            )
-        )
-    }
-    return dependencyList
-}
 
 @OptIn(ExperimentalTextApi::class)
 @Composable
